move array helper to utils.tsx
