+++ scripts/core/helpers/utils.tsx
@@ -117,3 +117,13 @@ export function copyString(data) {
     document.execCommand('copy');
     document.body.removeChild(element);
 }
+
+export function arrayMove<T>(arr: Array<T>, fromIndex: number, toIndex: number): Array<T> {
+    const copy = [...arr];
+    const element = copy[fromIndex];
+
+    copy.splice(fromIndex, 1);
+    copy.splice(toIndex, 0, element);
+
+    return copy;
+}
